* Refactoring of optional submodule locations
