+++ scistag/slidestag/flask/static/js/image_update.js
