+++ scistag/slidestag/flask/static/js/webcam_support.js
