+++ scistag/slidestag/kivy/__init__.py
@@ -1,0 +1,9 @@
+"""
+Integrates the Kivy specific extensions into SlideStag
+"""
+
+from .advanced_image import AdvancedImage
+from .embedded_slider import EmbeddedSlider
+from .simple_kivy_app import KivyApp
+
+__all__ = ["AdvancedImage", "EmbeddedSlider", "KivyApp"]
