+++ scistag/slidestag/kivy/advanced_image.py
