+++ scistag/slidestag/kivy/embedded_slider.py
