+++ scistag/slidestag/kivy/simple_kivy_app.py
