--- scistag/slidestag4kivy/__init__.py
@@ -1,3 +1,0 @@
-from .advanced_image import AdvancedImage
-
-__all__ = ["AdvancedImage"]
